requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
click-default-group==1.2.4
coloredlogs==15.0.1
cryptography==44.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
duckdb==1.1.3
durationpy==0.9
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
Faker==19.13.0
fastapi==0.115.8
filelock==3.16.1
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.38.0
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-experimental==0.3.4
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.1.147
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.64.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.12
overrides==7.7.0
pandas==1.5.3
pandasai==2.3.0
pillow==10.4.0
pluggy==1.5.0
posthog==3.15.1
propcache==0.2.1
protobuf==5.29.1
psutil==7.0.0
psycopg2==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pypdf==5.3.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySide6==6.8.0.2
pysqlite3-binary==0.5.4
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.4.1
shellingham==1.5.4
shiboken6==6.8.0.2
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlglot==25.33.0
sqlglotrs==0.3.0
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.45.3
streamlit==1.40.2
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.4.1
tqdm==4.67.1
transformers==4.49.0
triton==3.0.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.23
unstructured-client==0.30.3
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
xgboost==2.1.3
yarl==1.18.3
zipp==3.21.0
